Revert "Prohibit CREATE on static storages (ATTACH should be used instead)"
CREATE is used by Replicated database even for ATTACH, and anyway ATTACH
creates format_version.txt as well (without previuos patch).

This reverts commit 323f1f320ebe7e588d443abf78efa2c60193e7a9.

+++ src/Storages/StorageMergeTree.cpp
@@ -115,14 +115,12 @@ StorageMergeTree::StorageMergeTree(
     , writer(*this)
     , merger_mutator(*this)
 {
-    if (!attach && isStaticStorage())
-        throw Exception(ErrorCodes::TABLE_IS_READ_ONLY, "Creating data on static storage is prohibited, use ATTACH instead");
-
     initializeDirectoriesAndFormatVersion(relative_data_path_, attach, date_column_name);
+
 
     loadDataParts(has_force_restore_data_flag, std::nullopt);
 
-    if (!attach && !getDataPartsForInternalUsage().empty())
+    if (!attach && !getDataPartsForInternalUsage().empty() && !isStaticStorage())
         throw Exception(ErrorCodes::INCORRECT_DATA,
                         "Data directory for table already containing data parts - probably "
                         "it was unclean DROP table or manual intervention. "
